+++ server/routes/question/DeleteComment.js
@@ -10,20 +10,22 @@ async function DeleteComment(req, res, next) {
 
     if (!comment) return res.status(404).send(config.errorNotFound);
 
+    // Verify that the user owns the comment
     if (comment.creator !== user.username) {
         return res.status(403).send(config.errorForbidden);
     }
 
+    // Delete comment with BDPA server
     const { success } = await createRequest(
         'delete',
         `/questions/${questionID}/comments/${commentID}`
     );
 
+    if (!success) return res.status(500).send(config.errorGeneric);
+
     await Comment.findByIdAndDelete(commentID);
 
-    return success
-        ? res.sendStatus(200)
-        : res.status(500).send(config.errorGeneric);
+    return res.sendStatus(200);
 }
 
 module.exports = DeleteComment;
